Remove unused code; handle escaped and double-escaped script data; allow using buffer instead of string source for bench; enforce matching tags

+++ bench/build.sh
@@ -7,7 +7,11 @@ pushd "$(dirname "$0")"
 nodejs_cargo_toml="../nodejs/native/Cargo.toml"
 
 rm -rf node_modules
-cp "$nodejs_cargo_toml" "$nodejs_cargo_toml.orig"
+if [ -f "$nodejs_cargo_toml.orig" ]; then
+  echo 'Not altering Node.js Cargo.toml file'
+else
+  cp "$nodejs_cargo_toml" "$nodejs_cargo_toml.orig"
+fi
 sed -i 's%^hyperbuild = .*$%hyperbuild = { path = "../.." }%' "$nodejs_cargo_toml"
 HYPERBUILD_NODEJS_SKIP_BIN_DOWNLOAD=1 npm i
 mv "$nodejs_cargo_toml.orig" "$nodejs_cargo_toml"
